+++ ChangeLog.md
@@ -2,6 +2,7 @@
 
 ## Version v1.4.4 (not released yet)
 * Added support for parenthesis in row reference names.
+* Added support for reusable styles.
 
 ## Version v1.4.3 (submitted, not released yet)
 * Added additionalWidth property to make it possible to work with border style/width.